components folders & files reorganization
